Remove question all and add pages all
